pyproject.toml
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "adk-sql-reflection-pattern"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.17.0",
    "honcho>=2.0.0",
    "python-dotenv>=1.2.1",
    "toolbox-core>=0.5.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "poethepoet>=0.27.0",
    "pytest-asyncio>=1.2.0",
]
dataplex = [
    "Faker>=26.0.0",
    "pandas>=2.2.2",
    "google-cloud-dataplex>=1.14.0",
    "google-auth>=2.29.0",
    "google-cloud-resource-manager>=1.12.2",
]

[tool.setuptools]
packages = ["agents"]

[tool.poe.tasks]
datagen = "python scripts/generate_sample_data.py"
metadata = "python scripts/attach_metadata.py"
